Fix potential float/long casting but in interpolation
Reported by sonarcloud: https://sonarcloud.io/organizations/xenit-eu/rules\?open\=java%3AS2184\&rule_key\=java%3AS2184

+++ alfresco-health-processor-platform/src/main/java/eu/xenit/alfresco/healthprocessor/indexing/SimpleIndexingProgress.java
@@ -17,8 +17,9 @@ public class SimpleIndexingProgress implements IndexingProgress {
         this(startId, Instant.now(), endId, currentId);
     }
 
-    private static float interpolate(float start, float end, float current) {
-        return clampPercentage((current - start) / (end - start));
+    private static float interpolate(long start, long end, long current) {
+        // We are casting one side of the division to a float, so it performs float division instead of integer division
+        return clampPercentage((float)(current - start) / (end - start));
     }
 
     private static float clampPercentage(float in) {
